feat(onboard): make onboard page server first
