+++ src/app/onboard/loading.tsx
@@ -1,0 +1,6 @@
+import { Loading } from "@/components/Loading";
+import { TimerIcon } from "lucide-react";
+
+export default function LoadingPage() {
+  return <Loading message="Loading..." badge={TimerIcon} />;
+}
